v2.2.15: Sprint 4-5  Refactoring (-1382 LOC) + Performance optimizations
