+++ src/utils.js
@@ -15,6 +15,12 @@ const EXCLUDED_DIRS = ['node_modules', '.git', '.muaddib-cache'];
 let _extraExcludedDirs = [];
 let _scanRoot = '';
 
+/**
+ * Memoization cache for findFiles(). Key = dir|extensions|excludedDirs.
+ * Cleared between scans via clearFileListCache().
+ */
+const _fileListCache = new Map();
+
 function setExtraExcludes(dirs, scanRoot) {
   _extraExcludedDirs = Array.isArray(dirs) ? dirs : [];
   _scanRoot = scanRoot || '';
@@ -78,6 +84,21 @@ function findFiles(dir, options = {}) {
     depth = 0
   } = options;
 
+  // Top-level memoization: identical (dir, extensions, excludedDirs) → cached result
+  if (depth === 0) {
+    const cacheKey = dir + '|' + extensions.slice().sort().join(',') + '|' +
+      [...excludedDirs, ..._extraExcludedDirs].sort().join(',');
+    const cached = _fileListCache.get(cacheKey);
+    if (cached) return [...cached]; // return copy to prevent mutation
+    const result = _findFilesImpl(dir, { extensions, excludedDirs, maxDepth, results, visitedInodes, depth });
+    _fileListCache.set(cacheKey, [...result]);
+    return result;
+  }
+
+  return _findFilesImpl(dir, { extensions, excludedDirs, maxDepth, results, visitedInodes, depth });
+}
+
+function _findFilesImpl(dir, { extensions, excludedDirs, maxDepth, results, visitedInodes, depth }) {
   if (depth > maxDepth) return results;
   if (!fs.existsSync(dir)) return results;
 
@@ -114,7 +135,7 @@ function findFiles(dir, options = {}) {
           if (realStat.ino !== 0 && visitedInodes.has(realStat.ino)) continue;
           if (realStat.ino !== 0) visitedInodes.add(realStat.ino);
           if (realStat.isDirectory()) {
-            findFiles(realPath, { extensions, excludedDirs, maxDepth, results, visitedInodes, depth: depth + 1 });
+            _findFilesImpl(realPath, { extensions, excludedDirs, maxDepth, results, visitedInodes, depth: depth + 1 });
           } else if (extensions.some(ext => item.endsWith(ext))) {
             results.push(realPath);
           }
@@ -127,7 +148,7 @@ function findFiles(dir, options = {}) {
       if (lstat.ino !== 0) visitedInodes.add(lstat.ino);
 
       if (lstat.isDirectory()) {
-        findFiles(fullPath, { extensions, excludedDirs, maxDepth, results, visitedInodes, depth: depth + 1 });
+        _findFilesImpl(fullPath, { extensions, excludedDirs, maxDepth, results, visitedInodes, depth: depth + 1 });
       } else if (extensions.some(ext => item.endsWith(ext))) {
         results.push(fullPath);
       }
@@ -147,6 +168,10 @@ function findFiles(dir, options = {}) {
  */
 function findJsFiles(dir, results = []) {
   return findFiles(dir, { extensions: ['.js', '.mjs', '.cjs'], results });
+}
+
+function clearFileListCache() {
+  _fileListCache.clear();
 }
 
 /**
@@ -290,6 +315,7 @@ module.exports = {
   isDevFile,
   findFiles,
   findJsFiles,
+  clearFileListCache,
   escapeHtml,
   getCallName,
   Spinner,
