+++ vscode-extension/package.json
@@ -2,7 +2,7 @@
   "name": "muaddib-vscode",
   "displayName": "MUAD'DIB Security Scanner",
   "description": "Detecte les attaques supply chain npm directement dans VS Code",
-  "version": "2.2.14",
+  "version": "2.2.15",
   "publisher": "dnszlsk",
   "engines": {
     "vscode": "^1.85.0"
